refactor: add types to loader.ts
